Support mjs modules
This reshapes the source files to make index.mjs the source of truth and inlines the flow definitions using flow comments (removing the need for a separate file).

A new babel based build step replaces the mjs exports with cjs exports to produce an index.js artifact (which is no longer checked in) for distribution.